requirements.txt
streamlit
mysql-connector-python
pandas
streamlit-autorefresh